Handle merged images when fetching external comments
On rare occasion, a target comment's image would get merged with another, making the original image id in the comment link invalid.

When trying to fetch such a link, the site will simply redirect you back to the page that initiated the request.

The new code checks the response.url against the current page's, and uses the comments api to get the new image id then attempts fetching the comment preview again.